+++ ios/Classes/Modules/InstabugApi.m
@@ -5,6 +5,7 @@
 #import "IBGNetworkLogger+CP.h"
 #import "InstabugApi.h"
 #import "ArgsRegistry.h"
+#import "../Util/NativeUtils/IBGAPM+PrivateAPIs.h"
 
 #define UIColorFromRGB(rgbValue) [UIColor colorWithRed:((float)((rgbValue & 0xFF0000) >> 16)) / 255.0 green:((float)((rgbValue & 0xFF00) >> 8)) / 255.0 blue:((float)(rgbValue & 0xFF)) / 255.0 alpha:((float)((rgbValue & 0xFF000000) >> 24)) / 255.0];
 
@@ -279,14 +280,42 @@ extern void InitInstabugApi(id<FlutterBinaryMessenger> messenger) {
 
     NSString *gqlQueryName = nil;
     NSString *serverErrorMessage = nil;
+    NSNumber *isW3cHeaderFound = nil;
+    NSNumber *partialId = nil;
+    NSNumber *networkStartTimeInSeconds = nil;
+    NSString *w3CGeneratedHeader = nil;
+    NSString *w3CCaughtHeader = nil;
+    
     if (data[@"gqlQueryName"] != [NSNull null]) {
         gqlQueryName = data[@"gqlQueryName"];
     }
     if (data[@"serverErrorMessage"] != [NSNull null]) {
         serverErrorMessage = data[@"serverErrorMessage"];
     }
+    
+    if (data[@"partialId"] != [NSNull null]) {
+        partialId = data[@"partialId"];
+    }
+    
+    if (data[@"isW3cHeaderFound"] != [NSNull null]) {
+        isW3cHeaderFound = data[@"isW3cHeaderFound"];
+    }
+    
+    if (data[@"networkStartTimeInSeconds"] != [NSNull null]) {
+        networkStartTimeInSeconds = data[@"networkStartTimeInSeconds"];
+    }
+    
+    if (data[@"w3CGeneratedHeader"] != [NSNull null]) {
+        w3CGeneratedHeader = data[@"w3CGeneratedHeader"];
+    }
+    
+    if (data[@"w3CCaughtHeader"] != [NSNull null]) {
+        w3CCaughtHeader = data[@"w3CCaughtHeader"];
+    }
+    
+    
 
-    SEL networkLogSEL = NSSelectorFromString(@"addNetworkLogWithUrl:method:requestBody:requestBodySize:responseBody:responseBodySize:responseCode:requestHeaders:responseHeaders:contentType:errorDomain:errorCode:startTime:duration:gqlQueryName:serverErrorMessage:");
+    SEL networkLogSEL = NSSelectorFromString(@"addNetworkLogWithUrl:method:requestBody:requestBodySize:responseBody:responseBodySize:responseCode:requestHeaders:responseHeaders:contentType:errorDomain:errorCode:startTime:duration:gqlQueryName:serverErrorMessage:isW3cCaughted:partialID:timestamp:generatedW3CTraceparent:caughtedW3CTraceparent:");
 
     if ([[IBGNetworkLogger class] respondsToSelector:networkLogSEL]) {
         NSInvocation *inv = [NSInvocation invocationWithMethodSignature:[[IBGNetworkLogger class] methodSignatureForSelector:networkLogSEL]];
@@ -309,6 +338,11 @@ extern void InitInstabugApi(id<FlutterBinaryMessenger> messenger) {
         [inv setArgument:&(duration) atIndex:15];
         [inv setArgument:&(gqlQueryName) atIndex:16];
         [inv setArgument:&(serverErrorMessage) atIndex:17];
+        [inv setArgument:&(isW3cHeaderFound) atIndex:18];
+        [inv setArgument:&(partialId) atIndex:19];
+        [inv setArgument:&(networkStartTimeInSeconds) atIndex:20];
+        [inv setArgument:&(w3CGeneratedHeader) atIndex:21];
+        [inv setArgument:&(w3CCaughtHeader) atIndex:22];
 
         [inv invoke];
     }
@@ -316,6 +350,20 @@ extern void InitInstabugApi(id<FlutterBinaryMessenger> messenger) {
 
 - (void)willRedirectToStoreWithError:(FlutterError * _Nullable __autoreleasing *)error {
     [Instabug willRedirectToAppStore];
+}
+
+- (void)bindOnW3CFeatureFlagChangeCallbackWithError:(FlutterError * _Nullable __autoreleasing * _Nonnull)error {
+}
+
+
+- (nullable NSDictionary<NSString *,NSNumber *> *)isW3CFeatureFlagsEnabledWithError:(FlutterError * _Nullable __autoreleasing * _Nonnull)error {
+    NSDictionary<NSString * , NSNumber *> *result= @{
+        @"isW3cExternalTraceIDEnabled":[NSNumber numberWithBool:IBGNetworkLogger.w3ExternalTraceIDEnabled] ,
+        @"isW3cExternalGeneratedHeaderEnabled":[NSNumber numberWithBool:IBGNetworkLogger.w3ExternalGeneratedHeaderEnabled] ,
+        @"isW3cCaughtHeaderEnabled":[NSNumber numberWithBool:IBGNetworkLogger.w3CaughtHeaderEnabled] ,
+
+    };
+    return  result;
 }